refactor(warp): generation stages in place; wp select kernel; owned pipeline fully warp (M2 inc6)

+++ tests/test_warp_generate.py
@@ -28,7 +28,7 @@ def test_corner_count_sample(dev):
 
     # shape + dtype
     assert count.shape == (E,)
-    assert count.dtype == torch.long
+    assert count.dtype == torch.int32
     # range: every count in [min_num_points, max_num_points] inclusive
     assert int(count.min()) >= int(config.min_num_points)
     assert int(count.max()) <= int(config.max_num_points)
